requirements.txt
asgiref==3.4.1
astroid==2.3.3
autopep8==1.5
Babel==2.6.0
boto3==1.9.144
botocore==1.12.144
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==1.11.29
django-dotenv==1.4.2
django-forms-bootstrap==3.1.0
django-phonenumber-field==2.3.1
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
phonenumbers==8.10.10
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
six==1.14.0
sqlparse==0.4.3
stripe==2.24.1
typed-ast==1.4.1
typing_extensions==4.1.1
urllib3==1.24.2
wrapt==1.11.2
